Fix likes bug
