+++ src/index.js
@@ -1,5 +1,6 @@
 /* eslint-disable camelcase */
 import '../dist/output.css';
+import addLike from './modules/addLike';
 import getLikes from './modules/getLikes.js';
 import init from './modules/init.js';
 
@@ -25,30 +26,16 @@ init();
 
 // l1xzt8yaktrcb6minzZP
 
-// Add like to game using Involvement API and console.log the response
-const addLike = async (id) => {
-  fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/AvZcLNYpo6DEiCsZuxdL/likes/', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({
-      item_id: id,
-    }),
-  }).then(() => {
-    getLikes();
-  });
-};
-
-document.querySelector('#events').addEventListener('DOMSubtreeModified', () => {
-  document.querySelectorAll('svg').forEach((svg) => {
-    if (svg.id.includes('like')) {
-      svg.addEventListener('click', (e) => {
-        addLike(e.target.id);
-        e.target.style = 'fill: #f00';
+// Check if post exist before adding event listener to like button
+const check = () => {
+  const likeButtons = document.querySelectorAll('.like');
+  if (likeButtons) {
+    likeButtons.forEach((likeButton) => {
+      likeButton.addEventListener('click', () => {
+        console.log(likeButton.id);
       });
-    }
-  });
-});
-
-getLikes();
+    });
+  } else {
+    console.log('No like buttons found');
+  }
+};
